# openapi/pulse.yaml
openapi: 3.1.0
info:
  title: Pulse API
  version: 0.1.0
  description: |
    Pulse V2 — Personal Life OS backend.
    **Spec-first**: add every endpoint here before implementing it.
servers:
  - url: http://localhost:4000
    description: Local dev
  - url: https://api.pulse.app
    description: Production (Render)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/organizations:
    post:
      tags: [Organizations]
      summary: Create an organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:        { type: string }
                description: { type: string }
                avatar_url:  { type: string, format: uri }
      responses:
        '201': { description: Created }
    get:
      tags: [Organizations]
      summary: List my organizations
      responses:
        '200': { description: OK }
  /api/organizations/{id}:
    get:
      tags: [Organizations]
      summary: Organization detail + members
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: OK }
    patch:
      tags: [Organizations]
      summary: Update org (admin+)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:        { type: string }
                description: { type: string }
      responses:
        '200': { description: OK }
    delete:
      tags: [Organizations]
      summary: Delete org (owner)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Deleted }
  /api/organizations/{id}/events:
    get:
      tags: [Organizations]
      summary: List org events
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: OK }
    post:
      tags: [Organizations]
      summary: Create an org event (admin+)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, starts_at]
              properties:
                title:       { type: string }
                description: { type: string }
                starts_at:   { type: string, format: date-time }
                ends_at:     { type: string, format: date-time }
                location:    { type: string }
      responses:
        '201': { description: Created }
  /api/organizations/{id}/events/{eid}:
    patch:
      tags: [Organizations]
      summary: Update event (admin+)
      parameters:
        - { name: id,  in: path, required: true, schema: { type: string, format: uuid } }
        - { name: eid, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:       { type: string }
                description: { type: string }
                starts_at:   { type: string, format: date-time }
                ends_at:     { type: string, format: date-time }
                location:    { type: string }
      responses:
        '200': { description: OK }
    delete:
      tags: [Organizations]
      summary: Cancel event (admin+)
      parameters:
        - { name: id,  in: path, required: true, schema: { type: string, format: uuid } }
        - { name: eid, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Cancelled }
  /api/organizations/{id}/events/{eid}/rsvp:
    post:
      tags: [Organizations]
      summary: RSVP to an event
      parameters:
        - { name: id,  in: path, required: true, schema: { type: string, format: uuid } }
        - { name: eid, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [accepted, declined] }
      responses:
        '200': { description: OK }

  # ── Shares ──────────────────────────────────────────────────────────────────
...